refactor: move Order feat pages and components
