+++ src/features/Order/pages/OngoingOrders.tsx
@@ -9,9 +9,9 @@ import {
 } from "firebase/firestore";
 
 import { DeliveryStatusType } from "@/types/delivery.types";
-import Empty from "../../components/Empty";
-import { OrderConvert } from "../../converters/orders";
-import OrderItem from "./OrderItem";
+import Empty from "../../../components/Empty";
+import { OrderConvert } from "../../../converters/orders";
+import OrderItem from "../components/OrderItem";
 import { getAuth } from "firebase/auth";
 import { memo } from "react";
 import { useCollectionData } from "react-firebase-hooks/firestore";
