+++ src/features/Order/pages/Order.tsx
@@ -19,14 +19,14 @@ import { doc, getFirestore } from "firebase/firestore";
 import { useEffect, useState } from "react";
 import { useRecoilValue, useSetRecoilState } from "recoil";
 
-import AnimatedImg from "../../components/AnimatedImg";
-import { Branches } from "../../constants";
-import { OrderConvert } from "../../converters/orders";
-import OrderDescription from "../../utils";
+import AnimatedImg from "../../../components/AnimatedImg";
+import { Branches } from "../../../constants";
+import { OrderConvert } from "../../../converters/orders";
+import OrderDescription from "../../../utils";
 import { PaymentOptionType, PaymentStatusType } from "@/types/payment.types";
-import { orderAtom } from "../../atoms/order";
+import { orderAtom } from "../../../atoms/order";
 import { paymentGatewayURL } from "@/hooks/useCheckout";
-import { phpString } from "../../phpString";
+import { phpString } from "../../../phpString";
 import { useDocumentData } from "react-firebase-hooks/firestore";
 import { useParams } from "react-router";
 import { OrderType } from "@/types/cart.types";
